requirements.txt
fal-client>=0.5.0
python-dotenv>=1.0.0
gradio>=5.0.0
openai>=1.0.0
httpx>=0.27.0
mcp>=1.0.0
beautifulsoup4>=4.12.0
torch>=2.1.0
transformers>=4.40.0
accelerate>=0.27.0
librosa>=0.10.0
soundfile>=0.12.0
numpy>=1.26.0
Pillow>=10.0.0
